made some changes to authRoutes added role based authentication and rate limiting

+++ controllers/dyanmicController.js
@@ -1,4 +1,3 @@
-const { json } = require('express');
 const pool = require('../database/db.js')
 
 const dynamicController = async (req, res) => {
@@ -89,7 +88,9 @@ const dynamicController = async (req, res) => {
                     case 'interns':
 
                         const { intern_id: internId1, full_name, university, start_date, end_date, user_id } = column_values
-
+                        if (!full_name || !university || !start_date || !end_date || !user_id || !internId1) {
+                            return res.status(400).json({ message: 'Missing required fields for interns' });
+                        }
                         const internDataQuery = "SELECT * FROM Intern WHERE intern_id=$1"
                         const internValue = [internId1]
                         const internDataResult = await pool.query(internDataQuery, internValue)
@@ -115,7 +116,7 @@ const dynamicController = async (req, res) => {
                                     message: "Access Denied, Only Intern can update his info",
                                 });
                             }
-                        } else if(internData.full_name===full_name && internData.university===university) {
+                        } else if (internData.full_name === full_name && internData.university === university) {
                             if (role_id === 1) {//admin can update only start and end date and not interns personal info
                                 const updateQuery1 = "UPDATE Interns SET full_name=$1, university=$2, start_date=$3, end_date=$4 WHERE intern_id=$5 RETURNING *"
                                 const updateValues1 = [full_name, university, start_date, end_date, internId1]
@@ -131,7 +132,7 @@ const dynamicController = async (req, res) => {
                                     data: updateResult1.rows[0]
                                 });
                             }
-                            else{
+                            else {
                                 res.status(403).json({
                                     message: "Access Denied, Only admin  can update start and end date of internship info",
                                 });
@@ -139,26 +140,29 @@ const dynamicController = async (req, res) => {
                         }
                         break;
                     case 'internshipassignments':
-                        if(role_id===1){//only admin can update the internshipassignment details 
-                        const { assignment_id, intern_id: internId2, task_description, due_date } = column_values
-                        const updateQuery2 = "UPDATE InternshipAssignments SET task_description=$1, due_date=$2 WHERE assignment_id=$3 RETURNING *"
-                        const updateValues2 = [task_description, due_date, assignment_id]
-                        const updateResult2 = await pool.query(updateQuery2, updateValues2);
+                        if (role_id === 1) {//only admin can update the internshipassignment details 
+                            const { assignment_id, intern_id: internId2, task_description, due_date } = column_values
+                            if(!assignment_id || !internId2 ||!task_description||!due_date){
+                                return res.status(400).json({message: "Invalid Input Missing input fields"})
+                           }
+                            const updateQuery2 = "UPDATE InternshipAssignments SET task_description=$1, due_date=$2 WHERE assignment_id=$3 RETURNING *"
+                            const updateValues2 = [task_description, due_date, assignment_id]
+                            const updateResult2 = await pool.query(updateQuery2, updateValues2);
 
-                        if (updateResult2.rowCount == 0)
-                            return res.status(200).json({
-                                message: "Assignment Not Found",
+                            if (updateResult2.rowCount == 0)
+                                return res.status(200).json({
+                                    message: "Assignment Not Found",
+                                });
+
+                            res.status(200).json({
+                                message: "Data Updated Successfully",
+                                data: updateResult2.rows[0]
                             });
-
-                        res.status(200).json({
-                            message: "Data Updated Successfully",
-                            data: updateResult2.rows[0]
-                        });
-                    }else{
-                        res.status(403).json({
-                            message: "Access Denied, Only admin  can update internshipassignment info",
-                        });
-                    }
+                        } else {
+                            res.status(403).json({
+                                message: "Access Denied, Only admin  can update internshipassignment info",
+                            });
+                        }
                         break;
                     default:
                         res.status(400).json({ message: 'Invalid table name' });
@@ -169,43 +173,55 @@ const dynamicController = async (req, res) => {
                 // Handle delete operation
                 switch (table) {
                     case 'interns':
-                        const { intern_id: internId1 } = column_values;
-                        const deleteQuery1 = "DELETE FROM Interns WHERE intern_id=$1 RETURNING *";
-                        const deleteValues1 = [internId1];
-                        try {
+                        if (role_id === 1) {
+                            const { intern_id: internId1 } = column_values;
+                            const deleteQuery1 = "DELETE FROM Interns WHERE intern_id=$1 RETURNING *";
+                            const deleteValues1 = [internId1];
+                            try {
 
-                            if (!internId1) {
-                                throw new Error("Invalid input data");
+                                if (!internId1) {
+                                    throw new Error("Invalid input data");
+                                }
+                                const deleteResult1 = await pool.query(deleteQuery1, deleteValues1);
+
+                                if (deleteResult1.rowCount == 0)
+                                    throw new Error("Intern Not Found");
+
+                                res.status(200).json({
+                                    message: "Data Deleted Successfully",
+                                    data: deleteResult1.rows[0]
+                                });
+                            } catch (error) {
+                                res.status(404).json({ message: error.message });
                             }
-                            const deleteResult1 = await pool.query(deleteQuery1, deleteValues1);
-            
-                            if (deleteResult1.rowCount == 0)
-                                throw new Error("Intern Not Found");
-            
-                            res.status(200).json({
-                                message: "Data Deleted Successfully",
-                                data: deleteResult1.rows[0]
+                        } else {
+                            res.status(403).json({
+                                message: "Access Denied, Only admin  can perform delete operation",
                             });
-                        } catch (error) {
-                            res.status(404).json({ message: error.message });
                         }
                         break;
                     case 'internshipassignments':
-                        const { assignment_id } = column_values;
-                        const deleteQuery2 = "DELETE FROM InternshipAssignments WHERE assignment_id=$1 RETURNING *";
-                        const deleteValues2 = [assignment_id];
-                        try {
-                            const deleteResult2 = await pool.query(deleteQuery2, deleteValues2);
-            
-                            if (deleteResult2.rowCount == 0)
-                                throw new Error("Assignment Not Found");
-            
-                            res.status(200).json({
-                                message: "Data Deleted Successfully",
-                                data: deleteResult2.rows[0]
+                        if (role_id === 1) {
+                            const { assignment_id } = column_values;
+                            const deleteQuery2 = "DELETE FROM InternshipAssignments WHERE assignment_id=$1 RETURNING *";
+                            const deleteValues2 = [assignment_id];
+                            try {
+                                const deleteResult2 = await pool.query(deleteQuery2, deleteValues2);
+
+                                if (deleteResult2.rowCount == 0)
+                                    throw new Error("Assignment Not Found");
+
+                                res.status(200).json({
+                                    message: "Data Deleted Successfully",
+                                    data: deleteResult2.rows[0]
+                                });
+                            } catch (error) {
+                                res.status(404).json({ message: error.message });
+                            }
+                        } else {
+                            res.status(403).json({
+                                message: "Access Denied, Only admin  can delete internshipassignment info",
                             });
-                        } catch (error) {
-                            res.status(404).json({ message: error.message });
                         }
                         break;
                     default:
